Implemented Json with structure
This change refactors the postinst so that it fetches variables first
and stores them later. The variables can now contain `structure`
attribute which specifies sub-object it belongs to. It's a more powerful
version of renaming.

Additionaly, this change implements Json which already supports
structure. Other formats don't support structure yet.