Workaround for config ordering bug
This change implements a simple workaround for a case when the config
files are created in a different order than what they should be included
in. This leads to corruption of the configuration.

The temporary solution works for non-json files only, one level deep. It
achieves the goal by first writing all variable into the files and
including the files in another loop. Fortunately this is exactly what we
need now.

The proper solution has to sort the files as dependencies and process
them in that order.