Reverse the dependency logic of conf ext packages
Configuration extensions no longer depend on the extended packages, just
recommend them. While this seems a bit strange, there's a good reason
for that: if a service package is installed together with confext the
first time, the previous behavior causes restart of the service later,
while it'd be much better to prepare all configuration first and start
the service once after it's done.

Unfortunately, this may cause weird behavior on systems that don't
install recommends by default. In order to avoid issues, the packages
that require a service and a specific confext package now need to depend
on both, not just the confext package. Administrators need to resolve
this manually on their systems, but it's expected that they know what
they are doing if they avoided installing recommends in the first place.

While not a perfect solution to #28, it's preferrable to long install
times. The packages defining features can enforce ordering by having a
`no-features` package and depend on `no-features | feature` for each
known feature. Unfortunately, this introduces another empty package. In
some cases, when a package already depends on something else, the
dependency can be written as `soemthing-else | feature`. This looks
weird but seems to order the scripts correctly.

Closes #28