Patches removed from preinst
Calling dpkg-divert in preinst was a bad idea because we didn't replace
the original file right then. This caused packages to have broken files
during their postinst. We could of course patch it in preinst, but it'd
require us to include the patch file in the script, which is messy.

Instead we just remove patching from the preinst. It's handled in
postinst anyway, so there seems to be no good reason to do it in preinst.