Implemented migrations
This adds ability to specify `[migrations."<< VERSION"]` with `config`
set to a script that will run inside `configure` if the package is being
upgraded from a version prior to `VERSION`. The migrations are executed
from oldest to the newest so they will keep working across multiple
versions.

The `<<` operator is mandatory for readability and to allow adding more
complicated rules in the future.