Don't check migrations on invalid invokations
The `config` maintainer script is currently called with fixed action
`configure` and whatever second parameter was passed to `postinst`. As a
result, `config` is also called in cae of `triggered`, which causes dpkg
warnings about invalid versions. This would be fine so far as all
conditions fail, but may be confusing to the admin.

The change checks if a valid version was passed to the script. It's not
perfect and will obviously fail if a trigger is named such that it'd
look like a version but I don't think we could do any better and it's
not very probable anyway.