Depend on procps from service packages
Service packages use `systemd`, which internally uses `kill` to stop
processes, but doesn't depend on `procps`, which contains `kill`. This
is normally not a problem as Debian has `procps` installed by default
but breaks in certain edge cases like Docker images.

This change makes sure all service packages to depend on `procps`. We
don't know of any case when this would be too much implicitness, so it's
now hard-coded. It's possible to make it optional in the future if such
case is found.