Print the source file
Previously dep printing didn't include the source file since it shared
logic with writing `.d` files where printing it would be redundant since
to generate `.d` files one has to depend on it already. While it was
possible to workaround it in CI it was annoying and this simplifies it.