Implement import external files feature
When a project is built using debcrafter with detached specification
(external) additional files are sometimes required. So far this required
using something outside of debcrafter to handle this. While it worked,
it had several drawbacks:

* Duplication of code
* No validity checking (unless manually implemented)
* Adding the files to the list of source files was difficult
* It still required using awkward Debian syntax

And possibly more. This change adds a field that allows one to import
the files using debcrafter. The file is added to the dependencies as
well and the path is validated.