Add missing `#DEBHELPER#` to prerm
The generated prerm script was lacking the `#DEBHELPER#` marker which
prevented debhelper from filling in the scripts. Curiously, when `prerm`
was empty (which is fairly common) it would result in debhelper still
creating it and filling so it appeared as working in many cases. But
once the file was generated the package would broke.

This just adds the marker as the first step of removal process to ensure
removal of other components will not affect it.