Handle first-time installation of a database
When a package that previously didn't use a database starts using it
`dbconfig` gets confused thinking it's an upgrade from existing version
and doesn't attempt to initialize it. While it supports migration it
doesn't directly support migration from a without-database package to
with-database package. However it's possible to confuse it by not
passing the old package version to it so it thinks it's first-time
installation. This commit does it.